More coverage.
Add explicit # nocov to things we won't check, try to cover as much as we can. Also refactored and repaired some tests while I was at it.

The biggest change might be that I renamed the test files to match their associated R files. This makes it much easier to run tests and coverage for that single file. With proper mocking, the order of tests shouldn't matter.

I also updated to testtthat 3e for snapshots. At the same time, I change the capture.output tests to snapshots.